abie/requirements.txt
streamlit==1.32.0
langchain==0.1.16
langchain-openai==0.1.3
langchain-community==0.0.36
chromadb==0.4.24
openai==1.23.0
pypdf==4.1.0
pandas==2.2.1
matplotlib==3.8.4
python-dotenv==1.0.1
langchain-google-genai
